backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
sendgrid==6.10.0
